k8v-schema.pl

+++ build-scripts/20-configureOpenbox.sh
@@ -10,30 +10,36 @@ trap read debug  # puts a read request after each executable line
 ARCHBUILDS=~/ArchBuilds
     MACHINE=$ARCHBUILDS/sbMb
 
+#=> 0 ~/.config/openbox
+mkdir ~/.config/openbox
+
+#=> 0 openbox-themes
+cd ~/Arch/AUR
+git clone https://aur.archlinux.org/openbox-themes.git
+cd openbox-themes
+cat PKGBUILD | less
+makepkg -sic
+
+# Openbox Loma theme tweak for active window
+mkdir ~/.local/share/themes
+cp -r /usr/share/themes/Loma ~/.local/share/themes/LomaJH
+sed -i '/window.active.border.color:/ s/#000000/#FF8000/' ~/.local/share/themes/LomaJH/openbox-3/themerc
+sed -i '/^border.width:/ s/1/2/' ~/.local/share/themes/LomaJH/openbox-3/themerc
+
+#=> 1 importScreenshot.sh
+cp $ARCHBUILDS/jo/Openbox/openbox/importScreenshot.sh ~/.config/openbox/importScreenshot.sh
+chmod 755 ~/.config/openbox/importScreenshot.sh
+
 #=> 1 obmenu-generator
 cd ~/Arch/AUR
-sudo rm -r obmenu-generator
+[[ -d obmenu-generator ]] && sudo rm -r obmenu-generator
 git clone https://aur.archlinux.org/obmenu-generator.git
 cd obmenu-generator
 cat PKGBUILD | less
 makepkg -sic
-# now diff against the default
-nvim -O /etc/xdg/obmenu-generator/schema.pl $MACHINE/jo/openbox/schema.pl
-
-#=> 1 xinitrc
-ln -sf $MACHINE/jo/xinitrc ~/.xinitrc
-
-# #=> openbox-themes
-# cd ~/Arch/AUR
-# git clone https://aur.archlinux.org/openbox-themes.git
-# cd openbox-themes
-# cat PKGBUILD | less
-# makepkg -sic
-
-# # Openbox Loma theme tweak for active window
-# mkdir ~/.local/share/themes
-# cp -r /usr/share/themes/Loma ~/.local/share/themes/LomaJH
-# sed -i '/window.active.border.color:/ s/#000000/#FF8000/' ~/.local/share/themes/LomaJH/openbox-3/themerc
-# sed -i '/^border.width:/ s/1/2/' ~/.local/share/themes/LomaJH/openbox-3/themerc
-# cd ~
+if [[ -d $DROPBOX ]]; then # assume that my Dropbox is up
+    gvim -O /etc/xdg/obmenu-generator/schema.pl $MACHINE/jo/openbox/schema.pl
+else
+    cp /etc/xdg/obmenu-generator/schema.pl /mm/k8v-schema.pl # diff against this later
+fi
